switch supply field to double from float to fix precision related bug and add day advance animation

+++ Assets/Scripts/StrategicCombatCore/LandUnit.cs
@@ -11,8 +11,8 @@ namespace StrategicCombatCore
         public string objectId { get; set; }
         public GlobalString name = new();
         public int stregnth;
-        public float supplyTons;
-        public float supplyGeneratedTons; // Super Depot generate ~10,000 tons supply (Freight)
+        public double supplyTons;
+        public double supplyGeneratedTons; // Super Depot generate ~10,000 tons supply (Freight)
         public string remark;
 
         // public string strategicGroupId;
@@ -93,37 +93,15 @@ namespace StrategicCombatCore
             return 0;
         }
 
-        // public Cell cell
-        // {
-        //     get
-        //     {
-        //         var parentGroup = strategicGroupReference.Get();
-        //         if (parentGroup == null || !parentGroup.IsOnMap())
-        //             return null;
-        //         return parentGroup.cell;
-        //     }
-        // }
         public Cell cell => strategicGroupReference.GetCell();
 
-        // public SideState side
-        // {
-        //     get
-        //     {
-        //         var parentGroup = strategicGroupReference.Get();
-        //         if (parentGroup == null)
-        //             return null;
-        //         return parentGroup.side;
-        //     }
-        // }
         public SideState side => strategicGroupReference.GetSide();
 
         GlobalString ISupplyNetworkNode.GetName() => name;
-        // public void AddSupplyTons(float addedSupplyTons) => supplyTons += addedSupplyTons;
-        public float GetSupplyTons() => supplyTons;
-        public void SetSupplyTons(float value) => supplyTons = value;
+        public double GetSupplyTons() => supplyTons;
+        public void SetSupplyTons(double value) => supplyTons = value;
         public SupplyTransferState GetSupplyTransferState() => supplyTransferState;
         public bool IsDepotSameCellOnlySupply() => false;
-        // public LandUnit GetCurrentSourceDepot() => ((IStrategicGroupMemberReferenceable)this).GetCurrentSourceDepot();
     }
 }
